+++ app/components-react/shared/ButtonGroup.tsx
@@ -1,7 +1,7 @@
 import { Space } from 'antd';
 import React, { HTMLAttributes } from 'react';
 
-export function Buttons(p: HTMLAttributes<unknown>) {
+export function ButtonGroup(p: HTMLAttributes<unknown>) {
   return (
     <div style={{ textAlign: 'right', marginBottom: '8px' }}>
       <Space align="end">{p.children}</Space>
